+++ Demo_FreeRTOS/DEBUG_QEMU/Project_Settings/Startup_Code/Vector_Table.args
@@ -1,6 +1,6 @@
 -c
--I../RTD/include
 -I../FreeRTOS/Source/include
+-I../RTD/include
 -I../FreeRTOS/Source/portable/GCC/ARM_CM7/r0p1
 -I"C:/Users/vitoc/Desktop/workspace_group7/Demo_FreeRTOS/generate/include"
 -I"C:/Users/vitoc/Desktop/workspace_group7/Demo_FreeRTOS/RTD/include"
