--- Demo_FreeRTOS/src/Docs/FIX_SUMMARY.md
@@ -1,126 +1,0 @@
-# FreeRTOS Main.c Fix Summary - Key Issues Resolved
-
-## Problem Analysis from QEMU Output
-
-The original ELF was running but showing **"Transfer #X: FAILED! Errore nella verifica."** because:
-
-1. **QEMU Debug Output**: `lpspi_flush_txfifo: SPI transfer: tx=0x000000XX -> rx=0x00000000`
-
-    - Master was sending data correctly (TX values like 0xA1, 0xA2, etc.)
-    - **But RX was always 0x00000000** - slave wasn't responding properly
-
-2. **Timing Issue**: Race condition between slave preparation and master transfer start
-
-3. **Synchronization Problem**: Master starting before slave was fully armed/ready
-
-## Key Fixes Applied
-
-### 1. **Improved Timing and Synchronization**
-
-```c
-// BEFORE: No delay, potential race condition
-Lpspi_Ip_AsyncTransmit(&SLAVE_EXTERNAL_DEVICE, ...);
-xSemaphoreGive(producer_go);  // Immediate signal
-
-// AFTER: Proper timing control
-Lpspi_Ip_AsyncTransmit(&SLAVE_EXTERNAL_DEVICE, ...);
-vTaskDelay(pdMS_TO_TICKS(50));  // Wait for slave to be fully armed
-xSemaphoreGive(producer_go);    // Then signal master
-```
-
-### 2. **Enhanced Error Handling**
-
-```c
-// BEFORE: No timeout on semaphore operations
-xSemaphoreTake(producer_go, portMAX_DELAY);
-
-// AFTER: Proper timeout handling
-if (xSemaphoreTake(producer_go, pdMS_TO_TICKS(10000)) != pdTRUE) {
-    continue;  // Timeout recovery
-}
-```
-
-### 3. **Better Callback Safety**
-
-```c
-// BEFORE: Missing null check
-xSemaphoreGiveFromISR(slave_async_done_sem, &xHigherPriorityTaskWoken);
-
-// AFTER: Safe null check
-if (slave_async_done_sem != NULL) {
-    xSemaphoreGiveFromISR(slave_async_done_sem, &xHigherPriorityTaskWoken);
-}
-```
-
-### 4. **Improved Master Task Flow**
-
-```c
-// Added delay after slave ready signal
-if (xSemaphoreTake(producer_go, pdMS_TO_TICKS(10000)) != pdTRUE) {
-    continue;
-}
-vTaskDelay(pdMS_TO_TICKS(20));  // NEW: Extra safety delay
-
-// Added SPI status checking
-if (master_spi_status != LPSPI_IP_STATUS_SUCCESS) {
-    xSemaphoreGive(transfer_complete_sem);  // Continue cycle even on error
-    continue;
-}
-```
-
-### 5. **Enhanced Debug Output**
-
-```c
-// Added debug information for failed transfers
-sprintf(msg_buffer, "Debug: Master expected[0-2]: %02X %02X %02X, got: %02X %02X %02X\n",
-       slaveTxBuffer[0], slaveTxBuffer[1], slaveTxBuffer[2],
-       masterRxBuffer[0], masterRxBuffer[1], masterRxBuffer[2]);
-```
-
-### 6. **Robust Resource Management**
-
--   All semaphore creations now have error checking
--   Static variable declarations prevent memory corruption
--   Proper timeout values for all blocking operations
--   Graceful error recovery instead of infinite hangs
-
-## Expected Results
-
-The fixed code should resolve:
-
-1. **"Transfer FAILED" Messages**: Proper timing should allow slave to respond
-2. **Master RX = 0x00 Issue**: Slave should now send data back correctly
-3. **System Stability**: Better error handling prevents hangs
-4. **Debug Information**: More detailed output for troubleshooting
-
-## Technical Details
-
-### Root Cause
-
-The main issue was a **race condition** where:
-
-1. Slave task called `Lpspi_Ip_AsyncTransmit()`
-2. **Immediately** signaled master with `xSemaphoreGive(producer_go)`
-3. Master started `Lpspi_Ip_SyncTransmit()` **before slave was fully ready**
-4. Result: Master sent data but slave couldn't respond properly
-
-### Solution
-
-Added proper delays and sequencing:
-
-1. Slave arms itself with `Lpspi_Ip_AsyncTransmit()`
-2. **Wait 50ms** for slave to be fully ready
-3. **Then** signal master
-4. Master waits additional 20ms before starting transfer
-5. This ensures slave is completely prepared to respond
-
-## How to Test
-
-1. Rebuild project with new main.c
-2. Run with QEMU: `./qemu-system-arm -M nxps32k358evb -nographic -kernel Demo_FreeRTOS.elf -serial none -serial none -serial none -serial mon:stdio -d guest_errors`
-3. Look for:
-    - "SUCCESS! Dati verificati." messages instead of "FAILED!"
-    - Non-zero RX values in QEMU debug output
-    - Stable continuous operation
-
-The fix maintains all original FreeRTOS functionality while resolving the core timing/synchronization issues that were causing transfer failures.
