--- Demo_FreeRTOS/src/Docs/REFACTORING_SUMMARY.md
@@ -1,103 +1,0 @@
-# FreeRTOS Main.c Refactoring Summary
-
-## Key Improvements Made
-
-### 1. **Enhanced Task Synchronization**
-
--   **Before**: Used basic binary semaphores for all synchronization
--   **After**: Implemented a hybrid approach using:
-    -   Task notifications for lightweight signaling between tasks
-    -   Mutex for exclusive SPI bus access
-    -   Binary semaphores only where ISR communication is needed
-
-### 2. **Better Error Handling**
-
--   Added comprehensive error checking for all FreeRTOS API calls
--   Proper validation of semaphore/task creation
--   Timeout handling for all blocking operations
--   Graceful failure handling with debug messages
-
-### 3. **Improved Resource Management**
-
--   Used static variables to prevent stack corruption
--   Added proper mutex protection for SPI bus access
--   Increased task stack sizes to prevent stack overflow
--   Added system initialization flag to prevent premature operations
-
-### 4. **Enhanced Task Design**
-
--   **MasterTask**: Higher priority (3) for time-critical SPI operations
--   **SlaveTask**: Lower priority (2) for data verification and setup
--   Clear separation of concerns between tasks
--   Proper task state management using notifications
-
-### 5. **Robust Callback Implementation**
-
--   ISR-safe callback function using `xSemaphoreGiveFromISR`
--   Proper use of `portYIELD_FROM_ISR` for context switching
--   Dual notification system (semaphore + task notification)
-
-### 6. **Improved Debugging**
-
--   Centralized debug message function with safety checks
--   More detailed error reporting
--   Better status messages for troubleshooting
-
-## Code Structure Changes
-
-### Original Issues:
-
-1. No error checking for FreeRTOS object creation
-2. Potential race conditions in task synchronization
-3. Missing mutex protection for shared SPI resources
-4. No timeout handling for blocking operations
-5. Insufficient stack sizes
-6. Poor error recovery mechanisms
-
-### Fixed Implementation:
-
-```c
-// Example of improved error handling
-xResult = xTaskCreate(MasterTask, "MasterTask", MASTER_TASK_STACK_SIZE,
-                     NULL, MASTER_TASK_PRIORITY, &xMasterTaskHandle);
-if (xResult != pdPASS || xMasterTaskHandle == NULL) {
-    SendDebugMessage("ERROR: Failed to create Master Task\n");
-    return pdFAIL;
-}
-
-// Example of proper synchronization
-if (xSemaphoreTake(xSpiMutex, pdMS_TO_TICKS(1000)) != pdTRUE) {
-    SendDebugMessage("Master Task: Failed to acquire SPI mutex\n");
-    continue;
-}
-```
-
-## Key FreeRTOS Best Practices Implemented
-
-1. **Task Notifications**: Lightweight, fast signaling mechanism
-2. **Mutex Protection**: Prevents concurrent access to shared resources
-3. **Timeout Management**: All blocking calls have proper timeouts
-4. **Error Recovery**: System continues operation even on individual transfer failures
-5. **Resource Cleanup**: Proper initialization and error handling
-6. **Priority Management**: Appropriate task priorities for real-time operation
-
-## Expected Results
-
-The refactored code should:
-
--   Eliminate the "illegal instruction" error
--   Provide more robust SPI communication
--   Better handle error conditions
--   Improve overall system stability
--   Provide clearer debugging output
-
-## Next Steps
-
-To test the improvements:
-
-1. Rebuild the project using the proper S32DS IDE
-2. Flash the new firmware to the target hardware
-3. Monitor UART output for debug messages
-4. Verify successful SPI transfers between master and slave
-
-The new implementation follows FreeRTOS best practices and should resolve the runtime errors while providing a more maintainable and robust codebase.
